+++ pkg/tracing/span_index.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/uptrace/go-clickhouse/ch"
 	"github.com/uptrace/uptrace/pkg/tracing/xattr"
+	"github.com/uptrace/uptrace/pkg/tracing/xotel"
 )
 
 type SpanIndex struct {
@@ -60,7 +61,7 @@ func newSpanIndex(index *SpanIndex, span *Span) {
 	index.AttrKeys, index.AttrValues = attrKeysAndValues(span.Attrs)
 }
 
-func mapKeys(m AttrMap) []string {
+func mapKeys(m xotel.AttrMap) []string {
 	keys := make([]string, 0, len(m))
 	for key := range m {
 		keys = append(keys, key)
@@ -87,7 +88,7 @@ var (
 	indexedAttrSet = listToSet(indexedAttrs)
 )
 
-func attrKeysAndValues(m AttrMap) ([]string, []string) {
+func attrKeysAndValues(m xotel.AttrMap) ([]string, []string) {
 	keys := make([]string, 0, len(m))
 	values := make([]string, 0, len(m))
 	for k, v := range m {