+++ pkg/tracing/zipkin_handler.go
@@ -14,6 +14,7 @@ import (
 	"github.com/uptrace/bunrouter"
 	"github.com/uptrace/uptrace/pkg/bunapp"
 	"github.com/uptrace/uptrace/pkg/tracing/xattr"
+	"github.com/uptrace/uptrace/pkg/tracing/xotel"
 	"github.com/uptrace/uptrace/pkg/uuid"
 )
 
@@ -40,7 +41,7 @@ type ZipkinSpan struct {
 	Kind           string             `json:"kind"`
 	LocalEndpoint  ZipkinEndpoint     `json:"localEndpoint"`
 	RemoteEndpoint ZipkinEndpoint     `json:"remoteEndpoint"`
-	Tags           AttrMap            `json:"tags"`
+	Tags           xotel.AttrMap      `json:"tags"`
 	Annotations    []ZipkinAnnotation `json:"annotations"`
 }
 
@@ -111,7 +112,7 @@ func initSpanFromZipkin(dest *Span, src *ZipkinSpan) error {
 	dest.Attrs = src.Tags
 
 	if dest.Attrs == nil {
-		dest.Attrs = make(AttrMap)
+		dest.Attrs = make(xotel.AttrMap)
 	}
 	if src.LocalEndpoint.ServiceName != "" {
 		dest.Attrs.SetDefault(xattr.ServiceName, src.LocalEndpoint.ServiceName)